fix: special chars in title are now recognized
Special characters like the degree character are now recognized.

Also, the parser now recognizes questions with zero options before
trying to create requests for them and emits an error.

There was also a bug where NoCorrectIndicator error was not reported
as an error.

Form link is now copied to the clipboard if it was successfully created.

Also, if the title is empty, the parser now emits an error. The title
is the only thing in a form that is mandatory.